Schema and Data Ingest Method Changes
This patch changes the schema to include policy identifier and
modifies the result (pass/fail) to be an exit_code (int) as
REM and Compliance API do not know what a pass/fail should be.

This patch also combines the data ingestion methods from two
into a single method that uses form data to include report
data and files.

Signed-off-by: Ryan Brady <ryan.brady@anchore.com>